+++ test/pubdev_api_test.dart
@@ -1,6 +1,7 @@
 import 'package:pub_api_client/pub_api_client.dart';
 import 'package:test/test.dart';
 
+final packageName = 'fvm';
 void main() {
   group('PubDev Client', () {
     PubClient client;
@@ -10,7 +11,6 @@ void main() {
     });
 
     test('Can Fetch package info', () async {
-      final packageName = 'fvm';
       final packageInfo = await client.getPackage(packageName);
 
       final lastPubspec = packageInfo.latest.pubspec;
@@ -21,7 +21,6 @@ void main() {
     });
 
     test('Get package versions', () async {
-      final packageName = 'fvm';
       final packageInfo = await client.getPackage(packageName);
       final payload = await client.getVersions(packageName);
 
@@ -37,12 +36,35 @@ void main() {
     });
 
     test('Get Package Score', () async {
-      final payload = await client.getScore('fvm');
+      final payload = await client.getScore(packageName);
 
       expect(payload.lastUpdated, isNotNull);
       expect(payload.grantedPoints, isNotNull);
       expect(payload.likeCount, greaterThan(50));
       expect(payload.maxPoints, 110);
+    });
+
+    test('Check package update', () async {
+      final packageInfo = await client.getPackage(packageName);
+      final latestWithoutCurrent = await client.checkLatest(packageName);
+
+      expect(latestWithoutCurrent.needUpdate, false);
+      expect(latestWithoutCurrent.latestVersion, packageInfo.version);
+      expect(
+        latestWithoutCurrent.latestVersion,
+        packageInfo.versions.last.version,
+      );
+      expect(latestWithoutCurrent.packageInfo.name, packageName);
+
+      final latestNeedUpdate =
+          await client.checkLatest(packageName, currentVersion: '1.0.0');
+
+      expect(latestNeedUpdate.needUpdate, true);
+
+      final latestCurrent = await client.checkLatest(packageName,
+          currentVersion: packageInfo.version);
+
+      expect(latestCurrent.needUpdate, false);
     });
   });
 }