+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.4
+
+- Added check for package update
+- Added console printer for package update
+
 ## 1.0.3
 
 - Added user-agent to requests